submitted version (2023-11-24)
This is an unfortunately imperfect, unfinished version of my master's thesis, that I was obligated to hand in on due submission date, Nov 24th 2023. 
However, I will continue to work on this project, towards a full conceptualization and implementation of the mA³PAKE scheme, and its novel cryptographic primitives. 
Updates will follow, here.